fixed the routes because the pages werent loading properly

+++ public/js/delete-post.js
@@ -1,21 +1,19 @@
-const deletePost = async (e) => {
-    e.preventDefault(); 
-    const urlString = window.location.toString().split('/')
-    const postId  = urlString[4];
-    
-    const response = await fetch(`/api/post/${postId}`, {
-        method: 'DELETE'
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute("data-id")) {
+    const id = event.target.getAttribute("data-id");
+
+    const response = await fetch(`/api/post/${id}`, {
+      method: "DELETE",
     });
 
-    const data = await response.json()
     if (response.ok) {
-        alert('Post Deleted')
-        document.location.replace('/dashboard');
+      document.location.replace("/dashboard");
     } else {
-        alert("Something went wrong. Can't delete post")
+      alert("Something went wrong when deleting");
     }
-}
+  }
+};
 
-
-
-$('#delete-post').click(deletePost);
+document
+  .querySelector(".post-list")
+  .addEventListener("click", delButtonHandler);
